Add files via upload
This branch is a complete web interface test of One Time, Recurring, ACH and Customer Creation.  It uses a local .log file to assist debugging and a SQLExpress table to log request and responses.  This is not a complete application, but is designed to assist other users in getting up to speed quickly with PayTrace. Not supported by us or Paytrace but please enjoy.  

Created with VS2013.